change migration

+++ database/migrations/2023_04_19_080305_add_relationships_to_files_table.php
@@ -13,7 +13,7 @@ return new class extends Migration
     {
         Schema::table('files', function (Blueprint $table) {
             if (!Schema::hasColumn('files', 'folder_id')) {
-                $table->unsignedBigInteger('folder_id')->nullable();
+                $table->unsignedInteger('folder_id')->nullable();
                 $table->foreign('folder_id')->references('id')->on('folders');
             }
             if (!Schema::hasColumn('files', 'created_by_id')) {
